requirements.txt
streamlit
sentence-transformers
faiss-cpu
torch
groq
pypdf
numpy
scikit-learn
